requirements.txt
# PassForge - All-in-One Password Generator CLI
# Requirements

# Core Dependencies (Required for basic operation)
colorama>=0.4.6          # Cross-platform colored terminal output

# Feature Dependencies (Highly Recommended)
pyperclip>=1.8.2         # Clipboard support and auto-wipe functionality
zxcvbn>=4.4.28           # Advanced pattern analysis and strength checking
qrcode>=7.4.2            # QR code generation for OTP secrets (ASCII/Terminal)
cryptography>=42.0.0     # Secure encryption for local history logs

# Extension Dependencies (Optional)
pillow>=10.0.0           # Required for QR code IMAGE export (not needed for terminal)
pyyaml>=6.0              # Support for YAML configuration files

# Development & Testing
pytest>=8.0.0            # Unit testing framework
pytest-cov>=4.1.0        # Coverage reporting
pyinstaller>=6.0.0       # Building standalone executables

# PWA Dependencies (for web interface addon)
fastapi>=0.100.0         # Web framework for PWA backend
uvicorn>=0.23.0          # ASGI server for FastAPI
python-multipart>=0.0.6  # Support for form data in FastAPI
